Find Strongly Connected Components in directed graph C++
